Analysis/requirements.txt
pandas
numpy
matplotlib
pyserial
